whatpm/t/swml/tables-1.dat

#data
,
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
,a,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="2"
|             "a"

#data
,==,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="2"

#data
,a,==,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="3"
|             "a"

#data
,==,==,a,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="2"
|           <td>
|             colspan="2"
|             "a"

#data
,"a\"b","c\\d"
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"b"
|           <td>
|             "c\d"
